App/requirements.txt
streamlit
python-dotenv
torch
sentence-transformers
fastembed
qdrant-client
google-genai
openai